Rename event_sourcery.aggregate -> event_sourcery.event_sourcing
