--- event_sourcery/aggregate/__init__.py
@@ -1,7 +1,0 @@
-__all__ = [
-    "Aggregate",
-    "Repository",
-]
-
-from event_sourcery.aggregate.aggregate import Aggregate
-from event_sourcery.aggregate.repository import Repository
