+++ event_sourcery/event_sourcing/__init__.py
@@ -1,0 +1,7 @@
+__all__ = [
+    "Aggregate",
+    "Repository",
+]
+
+from event_sourcery.event_sourcing.aggregate import Aggregate
+from event_sourcery.event_sourcing.repository import Repository
