+++ event_sourcery/event_sourcing/aggregate.py
