+++ event_sourcery/event_sourcing/repository.py
@@ -2,7 +2,7 @@ from collections.abc import Iterator
 from contextlib import contextmanager
 from typing import Generic, TypeVar, cast
 
-from event_sourcery.aggregate import Aggregate
+from event_sourcery.event_sourcing import Aggregate
 from event_sourcery.event_store import (
     Event,
     EventStore,
