+++ tests/event_sourcing/test_aggregate.py
@@ -2,7 +2,7 @@ from uuid import uuid4
 
 import pytest
 
-from event_sourcery.aggregate import Aggregate, Repository
+from event_sourcery.event_sourcing import Aggregate, Repository
 from event_sourcery.event_store import Event, EventStore, StreamId, StreamUUID
 from event_sourcery.event_store.exceptions import ConcurrentStreamWriteError
